requirements.txt
google-genai==1.11.0
numpy
requests
pillow
